Optimized test execution time by separating TestCase with and without DB.
