+++ src/androidTest/java/com/couchbase/lite/BlobKeyTest.java
@@ -5,7 +5,7 @@ import com.couchbase.lite.util.Log;
 /**
  * Created by hideki on 2/20/15.
  */
-public class BlobKeyTest extends LiteTestCase{
+public class BlobKeyTest extends LiteTestCase {
     public void testConvertToHex(){
         String src = "Hello World!";
         Log.i(Log.TAG, "SRC => " + src);
