+++ src/androidTest/java/com/couchbase/lite/CRUDOperationsTest.java
@@ -25,7 +25,7 @@ import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 
-public class CRUDOperationsTest extends LiteTestCase implements Database.ChangeListener {
+public class CRUDOperationsTest extends LiteTestCaseWithDB implements Database.ChangeListener {
 
     public static final String TAG = "CRUDOperations";
