+++ src/androidTest/java/com/couchbase/lite/DatabaseTest.java
@@ -16,7 +16,7 @@ import java.util.concurrent.CountDownLatch;
 import java.util.concurrent.TimeUnit;
 import java.util.concurrent.atomic.AtomicInteger;
 
-public class DatabaseTest extends LiteTestCase {
+public class DatabaseTest extends LiteTestCaseWithDB {
 
     public void testPruneRevsToMaxDepthViaCompact() throws Exception {
