+++ src/androidTest/java/com/couchbase/lite/performance/Test10_DeleteDB.java
@@ -18,7 +18,7 @@
 package com.couchbase.lite.performance;
 
 import com.couchbase.lite.CouchbaseLiteException;
-import com.couchbase.lite.LiteTestCase;
+import com.couchbase.lite.LiteTestCaseWithDB;
 import com.couchbase.lite.Status;
 import com.couchbase.lite.TransactionalTask;
 import com.couchbase.lite.internal.Body;
@@ -28,7 +28,7 @@ import com.couchbase.lite.util.Log;
 import java.util.HashMap;
 import java.util.Map;
 
-public class Test10_DeleteDB extends LiteTestCase {
+public class Test10_DeleteDB extends LiteTestCaseWithDB {
 
     public static final String TAG = "DeleteDBPerformance";
