+++ src/androidTest/java/com/couchbase/lite/performance/Test13_QueryView.java
@@ -20,7 +20,7 @@ package com.couchbase.lite.performance;
 import com.couchbase.lite.CouchbaseLiteException;
 import com.couchbase.lite.Document;
 import com.couchbase.lite.Emitter;
-import com.couchbase.lite.LiteTestCase;
+import com.couchbase.lite.LiteTestCaseWithDB;
 import com.couchbase.lite.Mapper;
 import com.couchbase.lite.Query;
 import com.couchbase.lite.QueryEnumerator;
@@ -34,7 +34,7 @@ import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 
-public class Test13_QueryView extends LiteTestCase {
+public class Test13_QueryView extends LiteTestCaseWithDB {
 
     public static final String TAG = "QueryViewPerformance";
