+++ src/androidTest/java/com/couchbase/lite/performance/Test2_CreateDocsUnoptimizedWay.java
@@ -18,7 +18,7 @@
 package com.couchbase.lite.performance;
 
 import com.couchbase.lite.CouchbaseLiteException;
-import com.couchbase.lite.LiteTestCase;
+import com.couchbase.lite.LiteTestCaseWithDB;
 import com.couchbase.lite.Status;
 import com.couchbase.lite.internal.Body;
 import com.couchbase.lite.internal.RevisionInternal;
@@ -27,7 +27,7 @@ import com.couchbase.lite.util.Log;
 import java.util.HashMap;
 import java.util.Map;
 
-public class Test2_CreateDocsUnoptimizedWay extends LiteTestCase {
+public class Test2_CreateDocsUnoptimizedWay extends LiteTestCaseWithDB {
 
     public static final String TAG = "CreateDocsUnoptimizedWayPerformance";
