+++ src/androidTest/java/com/couchbase/lite/performance/Test8_DocRevisions.java
@@ -19,17 +19,14 @@ package com.couchbase.lite.performance;
 
 import com.couchbase.lite.CouchbaseLiteException;
 import com.couchbase.lite.Document;
-import com.couchbase.lite.LiteTestCase;
-import com.couchbase.lite.Status;
+import com.couchbase.lite.LiteTestCaseWithDB;
 import com.couchbase.lite.TransactionalTask;
-import com.couchbase.lite.internal.Body;
-import com.couchbase.lite.internal.RevisionInternal;
 import com.couchbase.lite.util.Log;
 
 import java.util.HashMap;
 import java.util.Map;
 
-public class Test8_DocRevisions extends LiteTestCase {
+public class Test8_DocRevisions extends LiteTestCaseWithDB {
 
     public static final String TAG = "DocRevisionsPerformance";
